requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
python-dotenv==1.0.1
imapclient==2.3.1
openpyxl==3.1.5
email-validator==2.2.0
apscheduler==3.10.4
